requirements.txt
# ============================================================
# Core ML/NLP Dependencies
# ============================================================
torch>=2.0.0
transformers>=4.36.0
datasets>=2.16.0
accelerate>=0.25.0
evaluate>=0.4.1
rouge-score>=0.1.2
sentencepiece>=0.1.99
sacrebleu>=2.4.0
nltk>=3.8.1

# ============================================================
# API & Web Framework
# ============================================================
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
pydantic>=2.5.0
python-multipart>=0.0.6
jinja2>=3.1.3

# ============================================================
# Data & Utilities
# ============================================================
pandas>=2.1.0
numpy>=1.24.0
python-box>=7.1.1
PyYAML>=6.0.1
tqdm>=4.66.0

# ============================================================
# Testing
# ============================================================
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.25.0

# ============================================================
# Development
# ============================================================
notebook>=7.0.0
ipywidgets>=8.1.0

# ============================================================
# Package
# ============================================================
-e .
